requirements.txt
fastapi==0.70.0
uvicorn[standard]==0.15.0
setuptools_scm
JPype1==1.3.0
sutime==1.0.1
spacy==3.2.0
matplotlib==3.5.0
fuzzywuzzy==0.18.0
pandas==1.3.4
word2number==1.1
python-dotenv==0.17.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
filelock==3.4.0
huggingface-hub==0.1.2
idna==3.3
joblib==1.1.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
six==1.16.0
tokenizers==0.10.3
torch==1.10.0
tqdm==4.62.3
transformers==4.12.5
typing_extensions==4.0.0
urllib3==1.26.7
sentencepiece==0.1.96
protobuf==3.19.1
pydantic==1.8.2
httpx==0.21.1
SQLAlchemy==1.4.27
psycopg2==2.9.2
databases==0.5.3
asyncpg==0.25.0
nltk==3.6.5
databases==0.5.3
asyncpg==0.25.0